requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
requests>=2.31
python-decouple>=3.8
google-generativeai>=0.8.0